finish added userService

+++ model/domain/User.go
@@ -5,5 +5,5 @@ type User struct {
 	Name     string `json:"name"`
 	Email    string `json:"email"`
 	Password string `json:"password"`
-	RoleId   int    `json:"role_id"`
+	RoleId   int64  `json:"role_id"`
 }
